+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update all depdencies (#124)
 - Fix multiarch docker image for arm64 users (#125)
 - Update markdown reference generator command to disable TOC (#127)
+- Add `am proxy` command (#128)
 
 ## [0.3.0]